sandlock-oci: run the in-sandbox PID-1 in-process instead of a static memfd init
