--- crates/sandlock-init/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "sandlock-init"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[lib]
-name = "sandlock_init"
-path = "src/proto.rs"
-
-[[bin]]
-name = "sandlock-init"
-path = "src/main.rs"
-
-[dependencies]
-libc = "0.2"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
