+++ crates/sandlock-oci/src/lib.rs
@@ -10,6 +10,7 @@
 //! - [`SupervisorCmd`] / [`SupervisorReply`] — IPC messages for the supervisor
 
 pub mod fdpass;
+pub mod init;
 pub mod policy;
 pub mod spec;
 pub mod state;
